feat(templates): add project links to VerticalFeatureRow
This commit adds project links to the VerticalFeatureRow component in the templates folder. Now, each feature row will have a "View Project" button that links to the respective GitHub repository. This enhances the user experience by providing easy access to the source code and allows users to explore the projects further.

The following