feat(doc): add documentation link to VerticalFeatureRow component
Add a new prop `docLink` to the `VerticalFeatureRow` component, allowing the user to provide a link to the documentation. This link will be displayed as a button labeled "阅读文档" (Read Documentation) and will open in a new tab when clicked. This change enhances the usability of the component by providing easy access to relevant documentation for each feature.