feat(templates): add i18n support to VerticalFeatures template
Add i18n support to the VerticalFeatures template by importing the `useTranslation` hook from "react-i18next" and using it to translate the title and description of the Section component. This allows for localization of the template content.

This commit also includes the necessary translation keys in the i18n configuration file for the new translated strings.